apiConfig/requirements.txt
python-dotenv==1.0.1
cohere==5.3.2
Flask==3.0.3